refactor : 레거시 레이어 정리(components, hooks, lib, templates)
