+++ app/src/features/flashcard/index.ts
@@ -1,6 +1,6 @@
 export { default as FlashCardPlayer } from './FlashCardPlayer';
 export type { FlashCardPlayerProps, DeleteMethod } from './FlashCardPlayer';
-export type { FlashCard } from './types';
+export type { FlashCard } from '@/entities/flashcard';
 export { useTodayFlashcards } from './model/useTodayFlashcards';
 export { useVisibilityDateCheck } from './model/useVisibilityDateCheck';
 export { generateDemoFlashcards } from './lib/demoFlashcards';
