+++ app/src/shared/ui/MarkdownBlock.tsx
@@ -1,1 +1,49 @@
-export { default } from '@/templates/MarkdownBlock';
+import React from 'react';
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+import { Light as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { github } from 'react-syntax-highlighter/dist/esm/styles/hljs';
+import 'github-markdown-css/github-markdown.css';
+
+interface MarkdownBlockProps {
+  markdown: string;
+}
+
+/**
+ * MarkdownBlock
+ * @deprecated - CodeDiffBlock 사용
+ */
+const MarkdownBlock: React.FC<MarkdownBlockProps> = ({ markdown }) => {
+  return (
+    <div className="markdown-body">
+      <ReactMarkdown
+        remarkPlugins={[remarkGfm]}
+        components={{
+          code({ inline, className, children, ...props }: any) {
+            const match = /language-(\w+)/.exec(className || '');
+            if (!inline && match) {
+              return (
+                <SyntaxHighlighter
+                  {...props}
+                  PreTag="div"
+                  children={String(children).replace(/\n$/, '')}
+                  language={match[1]}
+                  style={github}
+                />
+              );
+            }
+            return (
+              <code {...props} className={className}>
+                {children}
+              </code>
+            );
+          },
+        }}
+      >
+        {markdown}
+      </ReactMarkdown>
+    </div>
+  );
+};
+
+export default MarkdownBlock;
