+++ app/src/shared/ui/alert.tsx
@@ -1,1 +1,63 @@
-export * from '@/components/ui/alert';
+import * as React from 'react';
+import { cva, type VariantProps } from 'class-variance-authority';
+import { cn } from '@/shared/lib/cn';
+
+const alertVariants = cva(
+  'relative w-full rounded-lg border px-4 py-3.5 text-sm font-medium [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground [&>svg~*]:pl-7',
+  {
+    variants: {
+      variant: {
+        default: 'bg-card text-card-foreground border-border',
+        destructive:
+          'border-error/30 bg-error-bg text-error-text dark:border-error/30 dark:bg-error-bg dark:text-error-text',
+        success:
+          'border-primary/30 bg-success-bg text-success dark:border-primary/30 dark:bg-success-bg',
+      },
+    },
+    defaultVariants: {
+      variant: 'default',
+    },
+  }
+);
+
+const Alert = React.forwardRef<
+  HTMLDivElement,
+  React.HTMLAttributes<HTMLDivElement> & VariantProps<typeof alertVariants>
+>(({ className, variant, ...props }, ref) => (
+  <div
+    ref={ref}
+    role="alert"
+    className={cn(alertVariants({ variant }), className)}
+    data-slot="alert"
+    {...props}
+  />
+));
+Alert.displayName = 'Alert';
+
+const AlertTitle = React.forwardRef<
+  HTMLParagraphElement,
+  React.HTMLAttributes<HTMLHeadingElement>
+>(({ className, ...props }, ref) => (
+  <h5
+    ref={ref as React.Ref<HTMLParagraphElement>}
+    className={cn('mb-1 font-medium leading-none tracking-tight', className)}
+    data-slot="alert-title"
+    {...props}
+  />
+));
+AlertTitle.displayName = 'AlertTitle';
+
+const AlertDescription = React.forwardRef<
+  HTMLParagraphElement,
+  React.HTMLAttributes<HTMLParagraphElement>
+>(({ className, ...props }, ref) => (
+  <div
+    ref={ref as React.Ref<HTMLParagraphElement>}
+    className={cn('text-sm [&_p]:leading-relaxed', className)}
+    data-slot="alert-description"
+    {...props}
+  />
+));
+AlertDescription.displayName = 'AlertDescription';
+
+export { Alert, AlertTitle, AlertDescription };
