--- app/src/templates/MarkdownBlock.tsx
@@ -1,53 +1,0 @@
-import React from 'react';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import { Light as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { github } from 'react-syntax-highlighter/dist/esm/styles/hljs';
-import 'github-markdown-css/github-markdown.css';
-
-interface MarkdownBlockProps {
-  markdown: string;
-}
-
-/**
- * MarkdownBlock
- * @deprecated - CodeDiffBlock 사용
- * 
- * @param markdown - markdown 내용
- * @returns MarkdownBlock
- */
-const MarkdownBlock: React.FC<MarkdownBlockProps> = ({ markdown }) => {
-  return (
-    <div className="markdown-body">
-      <ReactMarkdown
-        remarkPlugins={[remarkGfm]}
-        components={{
-          code({ node, inline, className, children, ...props }: any) {
-            const match = /language-(\w+)/.exec(className || '');
-            if (match) {
-              return (
-                <SyntaxHighlighter
-                  {...props}
-                  PreTag="div"
-                  children={String(children).replace(/\n$/, '')}
-                  language={match[1]}
-                  style={github}
-                />
-              );
-            } else {
-              return (
-                <code {...props} className={className}>
-                  {children}
-                </code>
-              );
-            }
-          },
-        }}
-      >
-        {markdown}
-      </ReactMarkdown>
-    </div>
-  );
-};
-
-export default MarkdownBlock;
